# backend/api/src/main/python/openapi_server/controllers/animals_controller.py
import connexion
from typing import Dict
from typing import Tuple
from typing import Union

from openapi_server import util
from openapi_server.models.animal_config_update import AnimalConfigUpdate
from openapi_server.models.animal_update import AnimalUpdate
from openapi_server.models.animal_input import AnimalInput


def animal_config_get(animal_id):  # noqa: E501
    """Get animal configuration

    :param animal_id: 
    :type animal_id: strstr | bytes

    :rtype: Union[AnimalConfig, Tuple[AnimalConfig, int], Tuple[AnimalConfig, int, Dict[str, str]]]
    """
    # Auto-generated parameter handling

    # CMZ Auto-Generated Implementation Connection
    # This template automatically connects controllers to impl modules
    try:
        # Dynamic import of implementation module based on controller name
        # Auto-detect implementation module from operationId
        impl_module_name = "animalscontroller".replace("_controller", "")
        impl_function_name = "handle_"

        # Try common implementation patterns
        try:
            # Pattern 1: Direct module import
            impl_module = __import__(f"openapi_server.impl.{impl_module_name}", fromlist=[impl_function_name])
            impl_function = getattr(impl_module, impl_function_name)
        except (ImportError, AttributeError):
            # Pattern 2: Generic handler with hexagonal architecture routing
            from openapi_server.impl import handlers
            # Use the generic handle_ function that routes based on caller name
            impl_function = handlers.handle_
            if not impl_function:
                # Pattern 3: Default error for missing implementation
                raise NotImplementedError(f"Implementation function 'handle_' not found in handlers module")

        # Call implementation function with processed parameters
        result = impl_function(animal_id)

        # Handle different return types
        if isinstance(result, tuple):
            return result  # Already formatted (data, status_code)
        else:
            return result, 200

    except NotImplementedError as e:
        # Development mode: return clear error instead of placeholder
        from openapi_server.models.error import Error
        error_obj = Error(
            code="not_implemented",
            message=f"Controller animal_config_get implementation not found: {str(e)}",
            details={"controller": "AnimalsController", "operation": "animal_config_get"}
        )
        return error_obj, 501

    except Exception as e:
        # Use centralized error handler if available
        try:
            from openapi_server.impl.error_handler import handle_exception_for_controllers
            return handle_exception_for_controllers(e)
        except ImportError:
            # Fallback error response
            from openapi_server.models.error import Error
            error_obj = Error(
                code="internal_error",
                message=f"Internal server error in animal_config_get: {str(e)}",
                details={"controller": "AnimalsController", "operation": "animal_config_get"}
            )
            return error_obj, 500


def animal_config_patch(animal_id):  # noqa: E501
    """Update animal configuration

    :param animal_id: 
    :type animal_id: strstr | bytes

    :param animal_config_update: 
    :type animal_config_update:  | bytes

    :rtype: Union[AnimalConfig, Tuple[AnimalConfig, int], Tuple[AnimalConfig, int, Dict[str, str]]]
    """
    # Auto-generated parameter handling
    if connexion.request.is_json:
        animal_config_update = AnimalConfigUpdate.from_dict(connexion.request.get_json())  # noqa: E501

    # CMZ Auto-Generated Implementation Connection
    # This template automatically connects controllers to impl modules
    try:
        # Dynamic import of implementation module based on controller name
        # Auto-detect implementation module from operationId
        impl_module_name = "animalscontroller".replace("_controller", "")
        impl_function_name = "handle_"

        # Try common implementation patterns
        try:
            # Pattern 1: Direct module import
            impl_module = __import__(f"openapi_server.impl.{impl_module_name}", fromlist=[impl_function_name])
            impl_function = getattr(impl_module, impl_function_name)
        except (ImportError, AttributeError):
            # Pattern 2: Generic handler with hexagonal architecture routing
            from openapi_server.impl import handlers
            # Use the generic handle_ function that routes based on caller name
            impl_function = handlers.handle_
            if not impl_function:
                # Pattern 3: Default error for missing implementation
                raise NotImplementedError(f"Implementation function 'handle_' not found in handlers module")

        # Call implementation function with processed parameters
        result = impl_function(animal_id, animal_config_update)

        # Handle different return types
        if isinstance(result, tuple):
            return result  # Already formatted (data, status_code)
        else:
            return result, 200

    except NotImplementedError as e:
        # Development mode: return clear error instead of placeholder
        from openapi_server.models.error import Error
        error_obj = Error(
            code="not_implemented",
            message=f"Controller animal_config_patch implementation not found: {str(e)}",
            details={"controller": "AnimalsController", "operation": "animal_config_patch"}
        )
        return error_obj, 501

    except Exception as e:
        # Use centralized error handler if available
        try:
            from openapi_server.impl.error_handler import handle_exception_for_controllers
            return handle_exception_for_controllers(e)
        except ImportError:
            # Fallback error response
            from openapi_server.models.error import Error
            error_obj = Error(
                code="internal_error",
                message=f"Internal server error in animal_config_patch: {str(e)}",
                details={"controller": "AnimalsController", "operation": "animal_config_patch"}
            )
            return error_obj, 500


def animal_details_get(animal_id):  # noqa: E501
    """Fetch animal details

    :param animal_id: 
    :type animal_id: strstr | bytes

    :rtype: Union[AnimalDetails, Tuple[AnimalDetails, int], Tuple[AnimalDetails, int, Dict[str, str]]]
    """
    # Auto-generated parameter handling

    # CMZ Auto-Generated Implementation Connection
    # This template automatically connects controllers to impl modules
    try:
        # Dynamic import of implementation module based on controller name
        # Auto-detect implementation module from operationId
        impl_module_name = "animalscontroller".replace("_controller", "")
        impl_function_name = "handle_"

        # Try common implementation patterns
        try:
            # Pattern 1: Direct module import
            impl_module = __import__(f"openapi_server.impl.{impl_module_name}", fromlist=[impl_function_name])
            impl_function = getattr(impl_module, impl_function_name)
        except (ImportError, AttributeError):
            # Pattern 2: Generic handler with hexagonal architecture routing
            from openapi_server.impl import handlers
            # Use the generic handle_ function that routes based on caller name
            impl_function = handlers.handle_
            if not impl_function:
                # Pattern 3: Default error for missing implementation
                raise NotImplementedError(f"Implementation function 'handle_' not found in handlers module")

        # Call implementation function with processed parameters
        result = impl_function(animal_id)

        # Handle different return types
        if isinstance(result, tuple):
            return result  # Already formatted (data, status_code)
        else:
            return result, 200

    except NotImplementedError as e:
        # Development mode: return clear error instead of placeholder
        from openapi_server.models.error import Error
        error_obj = Error(
            code="not_implemented",
            message=f"Controller animal_details_get implementation not found: {str(e)}",
            details={"controller": "AnimalsController", "operation": "animal_details_get"}
        )
        return error_obj, 501

    except Exception as e:
        # Use centralized error handler if available
        try:
            from openapi_server.impl.error_handler import handle_exception_for_controllers
            return handle_exception_for_controllers(e)
        except ImportError:
            # Fallback error response
            from openapi_server.models.error import Error
            error_obj = Error(
                code="internal_error",
                message=f"Internal server error in animal_details_get: {str(e)}",
                details={"controller": "AnimalsController", "operation": "animal_details_get"}
            )
            return error_obj, 500


def animal_id_delete(id):  # noqa: E501
    """Delete an animal (soft delete)

    :param id: 
    :type id: strstr | bytes

    :rtype: Union[None, Tuple[None, int], Tuple[None, int, Dict[str, str]]]
    """
    # Auto-generated parameter handling

    # CMZ Auto-Generated Implementation Connection
    # This template automatically connects controllers to impl modules
    try:
        # Dynamic import of implementation module based on controller name
        # Auto-detect implementation module from operationId
        impl_module_name = "animalscontroller".replace("_controller", "")
        impl_function_name = "handle_"

        # Try common implementation patterns
        try:
            # Pattern 1: Direct module import
            impl_module = __import__(f"openapi_server.impl.{impl_module_name}", fromlist=[impl_function_name])
            impl_function = getattr(impl_module, impl_function_name)
        except (ImportError, AttributeError):
            # Pattern 2: Generic handler with hexagonal architecture routing
            from openapi_server.impl import handlers
            # Use the generic handle_ function that routes based on caller name
            impl_function = handlers.handle_
            if not impl_function:
                # Pattern 3: Default error for missing implementation
                raise NotImplementedError(f"Implementation function 'handle_' not found in handlers module")

        # Call implementation function with processed parameters
        result = impl_function(id)

        # Handle different return types
        if isinstance(result, tuple):
            return result  # Already formatted (data, status_code)
        else:
            return result, 204

    except NotImplementedError as e:
        # Development mode: return clear error instead of placeholder
        from openapi_server.models.error import Error
        error_obj = Error(
            code="not_implemented",
            message=f"Controller animal_id_delete implementation not found: {str(e)}",
            details={"controller": "AnimalsController", "operation": "animal_id_delete"}
        )
        return error_obj, 501

    except Exception as e:
        # Use centralized error handler if available
        try:
            from openapi_server.impl.error_handler import handle_exception_for_controllers
            return handle_exception_for_controllers(e)
        except ImportError:
            # Fallback error response
            from openapi_server.models.error import Error
            error_obj = Error(
                code="internal_error",
                message=f"Internal server error in animal_id_delete: {str(e)}",
                details={"controller": "AnimalsController", "operation": "animal_id_delete"}
            )
            return error_obj, 500


def animal_id_get(id):  # noqa: E501
    """Get a specific animal by ID

    :param id: 
    :type id: strstr | bytes

    :rtype: Union[Animal, Tuple[Animal, int], Tuple[Animal, int, Dict[str, str]]]
    """
    # Auto-generated parameter handling

    # CMZ Auto-Generated Implementation Connection
    # This template automatically connects controllers to impl modules
    try:
        # Dynamic import of implementation module based on controller name
        # Auto-detect implementation module from operationId
        impl_module_name = "animalscontroller".replace("_controller", "")
        impl_function_name = "handle_"

        # Try common implementation patterns
        try:
            # Pattern 1: Direct module import
            impl_module = __import__(f"openapi_server.impl.{impl_module_name}", fromlist=[impl_function_name])
            impl_function = getattr(impl_module, impl_function_name)
        except (ImportError, AttributeError):
            # Pattern 2: Generic handler with hexagonal architecture routing
            from openapi_server.impl import handlers
            # Use the generic handle_ function that routes based on caller name
            impl_function = handlers.handle_
            if not impl_function:
                # Pattern 3: Default error for missing implementation
                raise NotImplementedError(f"Implementation function 'handle_' not found in handlers module")

        # Call implementation function with processed parameters
        result = impl_function(id)

        # Handle different return types
        if isinstance(result, tuple):
            return result  # Already formatted (data, status_code)
        else:
            return result, 200

    except NotImplementedError as e:
        # Development mode: return clear error instead of placeholder
        from openapi_server.models.error import Error
        error_obj = Error(
            code="not_implemented",
            message=f"Controller animal_id_get implementation not found: {str(e)}",
            details={"controller": "AnimalsController", "operation": "animal_id_get"}
        )
        return error_obj, 501

    except Exception as e:
        # Use centralized error handler if available
        try:
            from openapi_server.impl.error_handler import handle_exception_for_controllers
            return handle_exception_for_controllers(e)
        except ImportError:
            # Fallback error response
            from openapi_server.models.error import Error
            error_obj = Error(
                code="internal_error",
                message=f"Internal server error in animal_id_get: {str(e)}",
                details={"controller": "AnimalsController", "operation": "animal_id_get"}
            )
            return error_obj, 500


def animal_id_put(id, animal_update):  # noqa: E501
    """Update an existing animal

    :param id: 
    :type id: strstr | bytes

    :param animal_update: 
    :type animal_update:  | bytes

    :rtype: Union[Animal, Tuple[Animal, int], Tuple[Animal, int, Dict[str, str]]]
    """
    # Auto-generated parameter handling
    if connexion.request.is_json:
        animal_update = AnimalUpdate.from_dict(connexion.request.get_json())  # noqa: E501

    # CMZ Auto-Generated Implementation Connection
    # This template automatically connects controllers to impl modules
    try:
        # Dynamic import of implementation module based on controller name
        # Auto-detect implementation module from operationId
        impl_module_name = "animalscontroller".replace("_controller", "")
        impl_function_name = "handle_"

        # Try common implementation patterns
        try:
            # Pattern 1: Direct module import
            impl_module = __import__(f"openapi_server.impl.{impl_module_name}", fromlist=[impl_function_name])
            impl_function = getattr(impl_module, impl_function_name)
        except (ImportError, AttributeError):
            # Pattern 2: Generic handler with hexagonal architecture routing
            from openapi_server.impl import handlers
            # Use the generic handle_ function that routes based on caller name
            impl_function = handlers.handle_
            if not impl_function:
                # Pattern 3: Default error for missing implementation
                raise NotImplementedError(f"Implementation function 'handle_' not found in handlers module")

        # Call implementation function with processed parameters
        result = impl_function(id, animal_update)

        # Handle different return types
        if isinstance(result, tuple):
            return result  # Already formatted (data, status_code)
        else:
            return result, 200

    except NotImplementedError as e:
        # Development mode: return clear error instead of placeholder
        from openapi_server.models.error import Error
        error_obj = Error(
            code="not_implemented",
            message=f"Controller animal_id_put implementation not found: {str(e)}",
            details={"controller": "AnimalsController", "operation": "animal_id_put"}
        )
        return error_obj, 501

    except Exception as e:
        # Use centralized error handler if available
        try:
            from openapi_server.impl.error_handler import handle_exception_for_controllers
            return handle_exception_for_controllers(e)
        except ImportError:
            # Fallback error response
            from openapi_server.models.error import Error
            error_obj = Error(
                code="internal_error",
                message=f"Internal server error in animal_id_put: {str(e)}",
                details={"controller": "AnimalsController", "operation": "animal_id_put"}
            )
            return error_obj, 500


def animal_list_get(status=None):  # noqa: E501
    """List animals

    :param status: Filter animals by status
    :type status: strstr | bytes

    :rtype: Union[List[Animal], Tuple[List[Animal], int], Tuple[List[Animal], int, Dict[str, str]]]
    """
    # Auto-generated parameter handling

    # CMZ Auto-Generated Implementation Connection
    # This template automatically connects controllers to impl modules
    try:
        # Dynamic import of implementation module based on controller name
        # Auto-detect implementation module from operationId
        impl_module_name = "animalscontroller".replace("_controller", "")
        impl_function_name = "handle_"

        # Try common implementation patterns
        try:
            # Pattern 1: Direct module import
            impl_module = __import__(f"openapi_server.impl.{impl_module_name}", fromlist=[impl_function_name])
            impl_function = getattr(impl_module, impl_function_name)
        except (ImportError, AttributeError):
            # Pattern 2: Generic handler with hexagonal architecture routing
            from openapi_server.impl import handlers
            # Use the generic handle_ function that routes based on caller name
            impl_function = handlers.handle_
            if not impl_function:
                # Pattern 3: Default error for missing implementation
                raise NotImplementedError(f"Implementation function 'handle_' not found in handlers module")

        # Call implementation function with processed parameters
        result = impl_function(status)

        # Handle different return types
        if isinstance(result, tuple):
            return result  # Already formatted (data, status_code)
        else:
            return result, 200

    except NotImplementedError as e:
        # Development mode: return clear error instead of placeholder
        from openapi_server.models.error import Error
        error_obj = Error(
            code="not_implemented",
            message=f"Controller animal_list_get implementation not found: {str(e)}",
            details={"controller": "AnimalsController", "operation": "animal_list_get"}
        )
        return error_obj, 501

    except Exception as e:
        # Use centralized error handler if available
        try:
            from openapi_server.impl.error_handler import handle_exception_for_controllers
            return handle_exception_for_controllers(e)
        except ImportError:
            # Fallback error response
            from openapi_server.models.error import Error
            error_obj = Error(
                code="internal_error",
                message=f"Internal server error in animal_list_get: {str(e)}",
                details={"controller": "AnimalsController", "operation": "animal_list_get"}
            )
            return error_obj, 500


def animal_post(animal_input):  # noqa: E501
    """Create a new animal

    :param animal_input: 
    :type animal_input:  | bytes

    :rtype: Union[Animal, Tuple[Animal, int], Tuple[Animal, int, Dict[str, str]]]
    """
    # Auto-generated parameter handling
    if connexion.request.is_json:
        animal_input = AnimalInput.from_dict(connexion.request.get_json())  # noqa: E501

    # CMZ Auto-Generated Implementation Connection
    # This template automatically connects controllers to impl modules
    try:
        # Dynamic import of implementation module based on controller name
        # Auto-detect implementation module from operationId
        impl_module_name = "animalscontroller".replace("_controller", "")
        impl_function_name = "handle_"

        # Try common implementation patterns
        try:
            # Pattern 1: Direct module import
            impl_module = __import__(f"openapi_server.impl.{impl_module_name}", fromlist=[impl_function_name])
            impl_function = getattr(impl_module, impl_function_name)
        except (ImportError, AttributeError):
            # Pattern 2: Generic handler with hexagonal architecture routing
            from openapi_server.impl import handlers
            # Use the generic handle_ function that routes based on caller name
            impl_function = handlers.handle_
            if not impl_function:
                # Pattern 3: Default error for missing implementation
                raise NotImplementedError(f"Implementation function 'handle_' not found in handlers module")

        # Call implementation function with processed parameters
        result = impl_function(animal_input)

        # Handle different return types
        if isinstance(result, tuple):
            return result  # Already formatted (data, status_code)
        else:
            return result, 201

    except NotImplementedError as e:
        # Development mode: return clear error instead of placeholder
        from openapi_server.models.error import Error
        error_obj = Error(
            code="not_implemented",
            message=f"Controller animal_post implementation not found: {str(e)}",
            details={"controller": "AnimalsController", "operation": "animal_post"}
        )
        return error_obj, 501

    except Exception as e:
        # Use centralized error handler if available
        try:
            from openapi_server.impl.error_handler import handle_exception_for_controllers
            return handle_exception_for_controllers(e)
        except ImportError:
            # Fallback error response
            from openapi_server.models.error import Error
            error_obj = Error(
                code="internal_error",
                message=f"Internal server error in animal_post: {str(e)}",
                details={"controller": "AnimalsController", "operation": "animal_post"}
            )
            return error_obj, 500