# backend/api/src/main/python/openapi_server/openapi/openapi.yaml
openapi: 3.0.3
info:
  description: API for the Cougar Mountain Zoo digital ambassador platform
  title: CMZ API
  version: 1.0.2
servers:
- description: Production
  url: https://api.cmz.example.com
- description: Staging
  url: https://staging.api.cmz.example.com
- description: Local
  url: http://localhost:8080
...
paths:
...
  /animal:
    post:
      operationId: animal_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnimalInput"
        required: true
      responses:
        "201":
          content:
            application/json:
              example:
                animalId: animal_123
                name: New Animal
                species: Test Species
                status: active
                created:
                  at: 2025-09-11T10:30:00Z
                modified:
                  at: 2025-09-11T10:30:00Z
                softDelete: false
              schema:
                $ref: "#/components/schemas/Animal"
          description: Animal created successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationError"
      security: []
      summary: Create a new animal
      tags:
      - Animals
      x-openapi-router-controller: openapi_server.controllers.animals_controller
  /animal/{id}:
    delete:
      operationId: animal_id_delete
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Animal deleted successfully
        "404":
          $ref: "#/components/responses/NotFound"
      security: []
      summary: Delete an animal (soft delete)
      tags:
      - Animals
      x-openapi-router-controller: openapi_server.controllers.animals_controller
    get:
      operationId: animal_id_get
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              example:
                animalId: animal_1
                name: Simba
                species: Lion
                status: active
                created:
                  at: 2025-08-11T09:00:00Z
                  by:
                    userId: user_1
                    email: system@cmz.org
                    displayName: System
                modified:
                  at: 2025-08-11T09:00:00Z
                  by:
                    userId: user_1
                    email: system@cmz.org
                    displayName: System
                softDelete: false
              schema:
                $ref: "#/components/schemas/Animal"
          description: Animal details
        "404":
          $ref: "#/components/responses/NotFound"
      summary: Get a specific animal by ID
      tags:
      - Animals
      x-openapi-router-controller: openapi_server.controllers.animals_controller
    put:
      operationId: animal_id_put
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnimalUpdate"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Animal"
          description: Animal updated successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
      security: []
      summary: Update an existing animal
      tags:
      - Animals
      x-openapi-router-controller: openapi_server.controllers.animals_controller
  /animal_config:
    get:
      operationId: animal_config_get
      parameters:
      - explode: true
        in: query
        name: animalId
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              example:
                voice: pokey_v1
                personality: goofy-friendly
                aiModel: gpt-5o-mini
                temperature: 0.7
                topP: 1
                toolsEnabled:
                - facts
                - media_lookup
                guardrails:
                  safe_mode: true
                created:
                  at: 2025-08-11T09:00:00Z
                  by:
                    userId: user_1
                    email: system@cmz.org
                    displayName: System
                modified:
                  at: 2025-08-11T09:00:00Z
                  by:
                    userId: user_1
                    email: system@cmz.org
                    displayName: System
                softDelete: false
              schema:
                $ref: "#/components/schemas/AnimalConfig"
          description: Animal config
      summary: Get animal configuration
      tags:
      - Animals
      x-openapi-router-controller: openapi_server.controllers.animals_controller
    patch:
      operationId: animal_config_patch
      parameters:
      - explode: true
        in: query
        name: animalId
        required: true
        schema:
          type: string
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnimalConfigUpdate"
        required: true
      responses:
        "200":
          content:
            application/json:
              example:
                animalConfigId: config_1
                voice: pokey_v1
                personality: goofy-friendly
                aiModel: gpt-5o-mini
                temperature: 0.7
                topP: 1
                toolsEnabled:
                - facts
                - media_lookup
                guardrails:
                  safe_mode: true
                created:
                  at: 2025-08-11T09:00:00Z
                  by:
                    userId: user_1
                    email: system@cmz.org
                    displayName: System
                modified:
                  at: 2025-08-11T09:00:00Z
                  by:
                    userId: user_1
                    email: system@cmz.org
                    displayName: System
                softDelete: false
              schema:
                $ref: "#/components/schemas/AnimalConfig"
          description: Updated config
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
      security: []
      summary: Update animal configuration
      tags:
      - Animals
      x-openapi-router-controller: openapi_server.controllers.animals_controller
  /animal_details:
    get:
      operationId: animal_details_get
      parameters:
      - explode: true
        in: query
        name: animalId
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              example:
                animalId: animal_1
                name: Pokey
                species: Coendou prehensilis
                status: active
                created:
                  at: 2025-08-11T09:00:00Z
                  by:
                    userId: user_1
                    email: system@cmz.org
                    displayName: System
                modified:
                  at: 2025-08-11T09:00:00Z
                  by:
                    userId: user_1
                    email: system@cmz.org
                    displayName: System
                softDelete: false
                description: A friendly prehensile-tailed porcupine ambassador.
                habitat: Tropical forests
                imageUrl: https://example.com/pokey.jpg
              schema:
                $ref: "#/components/schemas/AnimalDetails"
          description: Animal details
        "404":
          $ref: "#/components/responses/NotFound"
      summary: Fetch animal details
      tags:
      - Animals
      x-openapi-router-controller: openapi_server.controllers.animals_controller
...
components:
  responses:
...
    NotFound:
      content:
        application/json:
          examples:
            notFound:
              value:
                code: not_found
                message: Resource not found
          schema:
            $ref: "#/components/schemas/Error"
      description: Resource not found
    BadRequest:
      content:
        application/json:
          examples:
            badRequest:
              value:
                code: invalid_request
                message: The request is invalid
          schema:
            $ref: "#/components/schemas/Error"
      description: Bad request
    ValidationError:
      content:
        application/json:
          examples:
            validationError:
              value:
                code: validation_error
                message: Request validation failed
          schema:
            $ref: "#/components/schemas/Error"
      description: Validation error
  schemas:
...
    AuditFields:
      properties:
        created:
          $ref: "#/components/schemas/AuditStamp"
        modified:
          $ref: "#/components/schemas/AuditStamp"
        deleted:
          $ref: "#/components/schemas/AuditStamp"
        softDelete:
          default: false
          title: softDelete
          type: boolean
      title: AuditFields
      type: object
...
    Animal:
      allOf:
      - properties:
          animalId:
            type: string
          name:
            type: string
          species:
            type: string
          status:
            enum:
            - active
            - hidden
            type: string
        required:
        - animalId
        - name
        - species
        type: object
      - $ref: "#/components/schemas/AuditFields"
      example:
        deleted:
          at: 2000-01-23T04:56:07.000+00:00
          by:
            actorId: actorId
            displayName: displayName
            email: email
        species: species
        softDelete: false
        created:
          at: 2000-01-23T04:56:07.000+00:00
          by:
            actorId: actorId
            displayName: displayName
            email: email
        name: name
        modified:
          at: 2000-01-23T04:56:07.000+00:00
          by:
            actorId: actorId
            displayName: displayName
            email: email
        animalId: animalId
        status: active
      title: Animal
    AnimalInput:
      example:
        species: species
        name: name
        status: active
      properties:
        name:
          maxLength: 100
          minLength: 1
          title: name
          type: string
        species:
          maxLength: 100
          minLength: 1
          title: species
          type: string
        status:
          default: active
          enum:
          - active
          - hidden
          title: status
          type: string
      required:
      - name
      - species
      title: AnimalInput
      type: object
    AnimalUpdate:
      example:
        species: species
        name: name
        status: active
      properties:
        name:
          maxLength: 100
          minLength: 1
          title: name
          type: string
        species:
          maxLength: 100
          minLength: 1
          title: species
          type: string
        status:
          enum:
          - active
          - hidden
          title: status
          type: string
      title: AnimalUpdate
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
      x-bearerInfoFunc: openapi_server.controllers.security_controller.info_from_bearerAuth